+++ src/main/java/com/example/backend/controller/PinController.java
@@ -33,14 +33,11 @@ public class PinController {
 
     @PreAuthorize("hasAuthority('USER')")
     @RequestMapping("/pin/addPinByCoords")
-    public CommonResult addPinByCoords(@RequestBody Pin pin, @RequestParam(name = "id") Integer id) {
-        List<User> users = userService.findUserById(id);
-        if (users.size() == 0)
-            return CommonResult.failed("不存在id = " + id + "的user");
-        if (users.get(0).getType() == 0)
-            pin.setVisibility(0);
-        else if (users.get(0).getType() == 1)
-            pin.setVisibility(1);
+    public CommonResult addPinByCoords(@RequestBody Pin pin,
+                                       @RequestParam(name = "id") Integer id) {
+        User user = userService.findUserById(id).get(0);
+        // pin的可见性由后端设定
+        pin.setVisibility(user.getType());
         pin.setUser_id(id);
         Forum forum = new Forum();
         forumService.insertForum(forum);
@@ -54,12 +51,17 @@ public class PinController {
 
     @PreAuthorize("hasAuthority('USER')")
     @DeleteMapping("/pin/deletePin/{pin_id}")
-    public CommonResult deletePinById(@PathVariable(value = "pin_id", required = false) Integer pin_id) {
+    public CommonResult deletePinById(@PathVariable(value = "pin_id") Integer pin_id,
+                                      @RequestParam(value = "id") Integer id) {
         Pin pin = pinService.getPinById(pin_id);
         if (pin == null)
             return CommonResult.failed("不存在id = " + pin_id + "的pin");
         if (pin.getVisibility() == 1)
-            return CommonResult.failed("无法通过前端删除公共pin");
+            return CommonResult.failed("无法删除公共pin");
+
+        if (!pin.getUser_id().equals(id)) {
+            throw new RuntimeException("不要尝试删除他人的地图钉~");
+        }
 
         int photoRet = photoService.deletePhotoByPinId(pin_id);
         String msg = "";
@@ -113,16 +115,18 @@ public class PinController {
 
     @PreAuthorize("hasAuthority('USER')")
     @RequestMapping("/pin/changePinInfoById")
-    public CommonResult changePinInfoById(@RequestBody Pin pin, @RequestParam(name = "id") Integer id) {
-        List<User> users = userService.findUserById(id);
-        if (users.size() == 0)
-            return CommonResult.failed("不存在id = " + id + "的user");
-        if (users.get(0).getType() == 0)
-            pin.setVisibility(0);
-        else if (users.get(0).getType() == 1)
-            pin.setVisibility(1);
-        int ret = pinService.updatePin(pin);
-        if (ret == 1)
+    public CommonResult changePinInfoById(@RequestBody Pin pin,
+                                          @RequestParam(name = "id") Integer id) {
+        User user = userService.findUserById(id).get(0);
+
+        if (!pinService.getPinById(pin.getId()).getUser_id().equals(id)) {
+            throw new RuntimeException("不要尝试修改他人的地图钉~");
+        }
+
+        // pin的可见性由后端设定
+        pin.setVisibility(user.getType());
+
+        if (pinService.updatePin(pin) == 1)
             return CommonResult.success(null);
         else
             return CommonResult.failed("数据库不存在id = " + pin.getId() + "的pin");
@@ -174,9 +178,13 @@ public class PinController {
 
     @PreAuthorize("hasAuthority('USER')")
     @RequestMapping("/switchPos")
-    public CommonResult switchPos(@RequestBody Pin pin) {
-        int ret = pinService.switchPos(pin);
-        if (ret == 0)
+    public CommonResult switchPos(@RequestBody Pin pin,
+                                  @RequestParam(value = "id") Integer id) {
+        if (!pinService.getPinById(pin.getId()).getUser_id().equals(id)) {
+            throw new RuntimeException("不要尝试修改他人的地图钉~");
+        }
+
+        if (pinService.switchPos(pin) == 0)
             return CommonResult.failed("修改pin位置失败");
         else
             return CommonResult.success("修改pin位置成功");
